requirements.txt
pandas
numpy
scikit-learn
xgboost
rdkit-pypi
matplotlib
jupyter
